+++ backend/database.py
@@ -2,7 +2,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker, declarative_base
 
 # PostgreSQL 연결
-DATABASE_URL = "postgresql://postgres:6813@localhost:5432/repotify"
+DATABASE_URL = "postgresql://postgres:6813@localhost:5432/dump"
 
 engine = create_engine(DATABASE_URL)
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
